Rename PseudoTypeDynamic to DynamicPseudoType
Earlier we renamed all of the types to not have "Type" prefixes, but
missed this one because it actually starts with "PseudoType". Seems worth
retaining that explicit designation as a reminder to the caller that this
is not a real type, but this word ordering is more consistent with the
others.

Along with this, we rename the pseudo-type's only value to be
DynamicValue, which is less of a mouthful than PseudoTypeDynamicValue.