Start of value operations
The idea is that eventually this will grow into a well-tested library of
useful operations that other applications can build on either for their
own internal value wrangling or as a basis for an expression language.

For now, we just have Add and some other things that were needed to write
a reasonable unit test for Add.