Refactor Type to be struct rather than interface
Previously there was an asymmetry between Value and Type where the former
was a struct and the latter was an interface. This created an awkward
interface because all of the generic Type methods ended up being top-level
functions.

Instead now Type is a struct that embeds a private interface type for
the type implementations and can expose its own Type-generic methods for
more convenient access to Type tests and other operations.