Beginnings of a Set type
One of the target use-cases for this library has sets as a first-class
type. Robustly supporting sets in combination with all of the other
special cty features, like unknowns, is non-trivial. Thus having it built
in here seems beneficial so that calling applications can get a
well-tested implementation that is able to satisfy all of the relevant
guarantees.

This set type is not yet really functional. This is just stubbing out the
relevant types and establishing a child package to contain our low-level
set implementation.