Add NilVal value
NilVal is an invalid value that can be returned as a placeholder where
an error is also being returned. NilVal is an intentionally-broken value
that will cause panics if used as a value.

This is just provided here because using it provides more explicit meaning
than just returning directly the zero value of this type.