Traverse sets in a consistent order
Previously the ordering of set traversal was non-deterministic, which
is problematic for many situations but in particular for when one set is
nested inside another and thus we need to be able to generate a consistent
hash value.

Now the ordering is arbitrary but consistent for a given set of values.
Internally we just sort the bucket ids, so the values are returned sorted
by their CRC32 result and then insertion order if there are collisions.
Callers are not allowed to rely on this ordering, but they can be sure
that with a given version of this code they can iterate the same set
twice and get the same ordering.