Implement Value.Modulo
Since there's no built-in modulo for big.Float, we implement our own.
This is here primarily for use with positive integers, but since cty
Numbers are floats we also take a position on what it means to modulo
with non-positive or fractional numbers, which is a bit of a philosophical
matter (depends on how exactly one thinks of modulo) but the position
taken here is at least consistent and hopefully also defensible.