Value.HasIndex method
This allows easy checking of the key-based panic conditions in Index,
allowing a caller to avoid having to duplicate all of the relevant
validation rules. A language based on cty might also choose to expose
this as a function or operator so that users can take conditional actions
based on the presence of a particular key in a map.