convert: drop conversion to interface{} altogether for now
Conversion to interface{} was originally imagined as providing a default
set of mapping rules, similar to decoding a JSON structure into
interface{} with encoding/json .

However, the type systme of cty is more complex than JSON and so the
result is not likely to be convenient to use. This might come back later
should a concrete use-case emerge, but for now we'll just drop it.