Type.TestConformance method for type checking
This new method implements the idea of "type conformance", which allows
a concrete type to be matched against a type *specification*, where the
latter is a type that may be or include DynamicPseudoType in order to
permit any type.

The initial use-case for this is type checking of function arguments.