convert: wrap and unwrap with capsule types
Since capsule types encapsulate a specific native Go type, we can convert
to and from values of that type directly.

When unwrapping, if the target type is a pointer then the original pointer
can be recovered, thus giving access to the wrapped object itself rather
than to a copy. However, the caller is strongly recommended *not* to
mutate that object to preserve cty's usual assumption of value
immutability.