Support for Tuple types
Tuple types are similar to Object types except their "attributes" are
numeric indices rather than strings.

Tuples are supported primarily so that our type system is able to fully
represent all possible values resulting from decoding arbitrary JSON,
since otherwise we wouldn't be able to represent lists with mixed element
types.

From the caller's perspective they behave a lot like lists, except that
the index key must be known in order to determine the type and the
result of indexing can be of a different type depending on the index.

As of this commit they are supported in the core operations but not yet
supported in "convert" or in TestConformance.