convert: rename to gocty
Since this package converts between Go and Cty values, it ought to have
"go" in its name somewhere. Furthermore, we want to leave the name
"convert" free for the forthcoming package that will deal with converting
between different cty types in a standard, robust way so that this doesn't
need to be reinvented (likely with different behavior/semantics) in every
caller.