convert: Further stubbing of the public interface
Now includes the idea of type comparison and unification through
conversions, which will allow us to take an arbitrary set of types and
try to find a common supertype that they can all convert to. This will
be useful, for example, for down-converting objects into maps and tuples
into lists, where we need to transition from type-hetrogenous to
type-homogenous collections by finding a type all the elements have in
common, or for calling languages that are implementing conditional
evaluation constructs where both "sides" must produce the same type.