convert: Remove public CompareTypes function
This function is primarily here as a helper for the Unify function, and
it is implemented under the expectation that Unify will use it just as
a hint, verifying the result.

Thus we will make it private to avoid presenting the illusion that it is
intended to be generally useful.