convert: simplify different-kinded complex conversions
Since the compareTypes function is now internal it is free to make some
optimistic assumptions that wouldn't make sense for non-Unification
callers.

Here we say that we always want to try to promote sets to lists, sets
to tuples, lists to tuples and maps to objects, and then we'll let the
conversion logic itself figure out if such a thing is actually possible
given the actual element/attribute types.