convert: push DynamicPseudoType to the bottom of our preference list
Since DynamicPseudoType converts to anything -- under the optimistic
assumption that it will later get replaced with a concrete value of the
right type -- it should have the lowest preference. Our conversion logic
will do a just-in-time check of the real type once it's known, catching
any situations where a conversion is not actually possible.