Value.ElementIterator method
The Value.ForEachElement interface is rather inconvenient to use because
it prevents normal loop control flow, so instead we'll provide an iterator
interface that can be used with a for loop.

ForEachElement is retained as a wrapper around this, and for now
ElementIterator is tested only in terms of ForEachElement.

This also changes the way sets are iterated, with both the key and the
value set to the element for consistency with how we treat sets in
the Index and HasIndex methods.