function: Pass expected return type to impl callback
Implementation experience with generic HIL functions suggests that it's
convenient to pass in the result of the return type function to the
implementation function so that generic functions (functions whose return
type depends on the arguments) don't need to re-implement their
type-selection logic in the implementation function.

This is particularly useful if, for example, the function is using
type unification across a list of dynamic varargs; unification can be
expensive, so it's good to do that only once during the type check and
re-use that result in the implementation.