function/stdlib: establish package of "standard library" functions
The idea with this is that applications building on cty ought to behave
in a somewhat-consistent way with regard to certain fundamental functions,
including consistent handling of unicode strings, big floats, etc. By
centralizing these here it's more likely that their implementations will
be common across multiple callers, particularly if these functions are
used as part of the implementation of other functions.