Remove "storable" stubs
This was an earlier idea of how JSON serialization (and serialization in
general) might be handled, but we're now going to go in a different
direction by providing specialized support for each serialization format
we need. We already have Gob support, and JSON support will follow
shortly in a separate package.