json: ImpliedType function
ImpliedType determines the cty type that can be assumed from the structure
of the given JSON buffer.

This uses the most general cty type for each JSON type, ensuring that
the value can be decoded without loss. However, it can also produce
results that may be inconvenient for some callers to handle, such as
DynamicPseudoType for null, and so most callers should dictate a specific
type they want and have other functions in this package (still to be
written) ensure conformance with it.