Remove SetElemStep
A convention has emerged that a set behaves as a collection whose keys
and values are both the set elements, with a set element acting as its
own key.

Thus we don't actually need a specialized PathStep for set elements: we
can just use the IndexStep with the Key set to the element being
traversed.