json: Rework how Unmarshal is implemented
Previously this was done at a rather low level using raw JSON tokens.
With fresh eyes I see that there's a better way to write this using the
higher-level features of the JSON decoder, preferring to deal with JSON
buffers as the primary currency rather than decoders.

This simplifies things considerably since now the JSON package does more
of the work and we don't need to deal with the weirdness of already having
read a token before we know what we wanted to parse.