function: suppress panics from function implementations
The function abstraction is a convenient place to intercept panics and
promote them to errors, thus freeing the implementers of languages that
build on cty from having to handle these panics.

A panic in a function is still considered a bug in that function, but
this way such a bug is isolated to just the invocation of that function
and can't take down the entire calling application.