function/stdlib: catch and suppress ErrNaN
The arithmetic functions on big floats (which is the internal
implementation of the cty arithmetic operations) panics with ErrNaN any
time a NaN would be returned in conventional floating point arithmetic.

Since we deal in error messages rather than panics, and we want to return
the error at a cty level of abstraction rather than Go, here we intercept
those panics and turn them into friendlier errors.