path.Copy method
We often require callers to create their own copies of paths received in
callbacks, etc, because while we're walking a data structure we just
mutate the same path in-place to avoid too much allocation.

This method gives callers a convenient way to create a copy if they need
to have a path whose validity lifetime they control.