diff: shallow list diffing
This produces a sequence of deletes, inserts and contexts for the direct
values of a pair of two lists of the same type. It does this by computing
one of the longest common sub-sequences of the two lists and then
walking all three lists (old, new, longest common) to produce the
delete, insert and context items in the diff.