convert: fix logic error in converting lists to lists/sets
Previously this codepath had a problem whereby it would still return a
conversion even if the two element types were incompatible, and that
conversion would then be a no-op on the list elements, allowing the
unconverted and incorrectly typed list elements to pass through.

Now we recognize the difference between the two collections having the
same element type and the two being unconvertable, returning nil in the
latter case to signal to the caller that the outer type is therefore
unconvertable too.