stdlib: Use textseg for string operations
Previously we were abusing a side-effect of the Unicode normalization
library, which is able to recognize certain grapheme clusters due to its
job of normalizing the encoding of some that can be written in multiple
ways.

Now we're using a library that implements the grapheme splitting rules
defined in Unicode TR#29, which is true to what the doc comments were
claiming all along.