cty: fix edge-cases in RawEquals method
The RawEquals method was previously trying to just delegate to Equals
and catch a few things it wouldn't support, but that is not suitable
because there can be unknowns and dynamics inside nested structures which
still need to be handled.

Instead, then, we'll actually recurse into the structure using RawEquals
and do our raw comparison at each level of the structure.