UnknownAsNull helper function
In some cases it's necessary to serialize a value that might contain
unknown values in a format that doesn't support them, such as JSON. As
long as the unknown-ness of these values is not actually important, this
helper function will turn them into nulls of the same type which can then
be serialized.

It is not possible to recover the unknown values later, so this should be
used with care only in situations where unknown values would not be
handled anyway, such as type checking.