function/stdlib: Format function
This is a cut down version of "printf" that makes sense for the cty type
system. It does not have all of the features of the C or Go
implementations but is roughly a subset of Go's fmt implementation.

For consistency with how our other library functions behave, it uses
grapheme clusters as its definition of "character" for string
manipulation.

Also for consistency with our usual behavior, errors are produced where
impossible situations are requested, which is in contrast with the Go
implementation where placeholder sequences are instead inserted directly
into the result. cty functions don't use this sort of in-band signaling,
instead preferring to fail explicitly so the user can quickly notice a
mistake and correct it.

The main guts of this function are a finite state machine implemented
using Ragel. As a conseqeunce, ragel must be installed in order to run
"go generate" on this package.