cty: Value.IsWhollyKnown
This is similar to Value.IsKnown but it works recursively into structural
and collection types to find any nested unknown values.

This is useful in some isolated cases where the result of an operation
depends on the entire nested structure. In particular, it's important for
JSON serialization since the entire structure must be known in order to
produce a JSON version of it.