function/stdlib: JSONEncode correctly handles nested unknowns
Previously we were only dealing with shallow unknowns, and so a structure
or collection containing a nested unknown value would cause an error.

Now we detect this and return our result as unknown, which correctly
reflects that we should be able to produce a known string once the
unknown value is replaced with a known value.