cty: NormalizeString function
cty.StringVal applies NFC normalization to incoming strings, which means
that if they are later extracted using Value.AsString then they may not
be byte-for-byte identical to what was originally provided.

Exporting NormalizeString gives callers a way to normalize a native Go
string for comparison without all of the ceremony of wrapping and then
immediately unwrapping with cty.StringVal.