cty: NFC-normalize map keys and object attribute names
We've always normalized the strings encapsulated in the cty.String type,
but we didn't do this normalization for map keys or object attribute names.

Doing this for map keys is important because we index into maps using
cty.String values, and so an un-normalized key would never be indexable.

It's less significant for object attributes, but our conversion rules
(in the cty/convert package) assume that object attributes can be turned
into map keys losslessly and vice-versa, so we also apply that
normalization to attributes to preserve this assumption.