cty: Walk and Transform functions
Implementation experience so far has suggested that there's a common need
for both visiting all of the nodes in a cty data structure for inspection
and for performing "surgical" transformations on a value while leaving
the general structure intact.

These two new helper functions handle the mechanics of recursing into
all of the different data structures in the correct way and leave the
caller just to implement the per-value logic.