convert: Conversion between object types
cty is intended to be a structural type system on object attributes, and
so object type B is a subtype of object type A if B contains at least
all of the attributes of A.

Consistent with the type conversion rules elsewhere, the attribute types
do not need to match exactly but must have a recursive conversion
available.

To better handle cases where object-to-object conversions fail, the new
helper function MismatchMessage can produce a more helpful
English-language description of a type mismatch between objects, where
before Convert would simply have reported "an object is required".