cty: Type.FriendlyNameForConstraint method
When a cty.Type is representing a type _constraint_, the value
cty.DynamicPseudoType represents "any type", and so it was previously
confusing when a non-conformance error would report something like
"set of dynamic required".

FriendlyNameForConstraint is a variant of FriendlyName that refers to
DynamicPseudoType as "any type". It also has a special case of collections
of dynamic which report as "of any single type", to try to make it
clearer that all of the element types must match even though any type is
allowed.