convert: Specialized mismatch messages for structure-to-collection
When converting from structure to collection with a dynamic element type,
a conversion failure results if we can't find a single type to convert
all of the structure elements/attributes to.

Here we return some specialized error messages in such cases, to help
users understand what needs to be corrected.