cty: New PathSet type
This began its life in a prototype branch in Terraform, but it seems like
passing around sets of paths is a common need for capturing out-of-band
characteristics of certain paths within an object, such as whether it
changed compared to some other object, etc.

This builds on the same set utility type we use for cty.Set values, but
with a different "rules" implementation that is better suited to sets of
paths.