cty/convert: Conversion to DynamicPseudoType should pass through verbatim
The intent of a conversion to DynamicPseudoType is to say "no conversion
required", for convenience in situations where a caller is constructing
conversion requirements from its input, to avoid the need for the caller
_itself_ to handle this special case.

We were correctly handling this for null, but not for unknown. Now we'll
retain the type of unknown values through conversion to dynamic too, so
that callers are able to continue with their type inference/checking.