cty/convert: Better MismatchMessage for collection/structural types
When we're comparing two compound types of compatible kind, it's better to
make a statement about the element types rather than the top-level types
in order to give the user better feedback on how to address the problem.

This also includes some regression tests specifically for MismatchMessage.
Previously it was tested only indirectly in other tests as part of error
messages.