cty/convert: Allow unification of object/tuple to produce new type
Our normal behavior for type unification is to select the most general of
the given types, but when structural type kinds are involved we can get
better outcomes by potentially constructing an entirely new type that is
the unification of all of the corresponding structure elements, where
possible.

For example, when trying to find a common type between a tuple with a
single string and a tuple with two strings, it is better to unify to a
list of string rather than to fail unification altogether.

This also includes a new conversion from one tuple type to another where
both have the same number of elements and convertible types. This
compliments the object-to-object conversion and allows potentially-richer
unification outcomes with tuples where the tuple-ness (and thus the
distinct element types) can be preserved if the number of elements
matches.