cty: Centralize string-to-number conversion
It's important to use a consistent precision and rounding rule when
parsing numbers that do not have exact representations in binary, or else
we end up with numbers that are not equal even though they started from
the same string.

Here we use the new function consistently in all of the internal
conversions. Any calling application accepting numbers as strings should
also use this function, to ensure consistency.