cty/convert: Avoid deep recursion when object-to-map unification fails
Due to some over-zealous copying from an earlier function, there was a
codepath in unifyObjectTypesToMap which would recursively call itself with
exactly the same arguments, leading to infinite recursion.

A comment here suggested that this codepath should be unreachable, but
we've seen this be reached at least once (sadly, I was unable to reproduce
it in such a way to let me write a test for it) and so if we hit that
codepath we'll prefer to indicate that no conversion is possible rather
than potentially causing a stack overflow.

With this panic fixed, hopefully if this weird situation is found again in
future it'll be easier to collect information about what situation led to
it, since the end-user that receives it will get a better contextual
error.