cty: Comprehensive implementation of fmt.GoStringer on Value
Previously we were lacking handling of some of the cty types in this
method, causing hard-to-read debug output in calling applications.

This new implementation fully covers the space of values, so we should
no longer see the implementation-details-exposing default result.

This also includes a new function MustParseNumberVal, which makes it
easier to construct number values that don't fit in int64 or float64 and
also gives us a better representation to use in the GoString results for
big numbers.