function/stdlib: new "range" function
This generates sequences of numbers using a start, and end, and a step
value.

This is primarily here so that calling applications that support iterating
over sequences can generate small sequences of numbers to iterate over.
To avoid bad arguments causing huge memory usage, the number of results
is artificially limited to 1024, after which the function returns an
error.