Unify dynamic with object or tuple types
Unify already will return a primitive type when a dynamic type exists in
the list. When presented with objects or tuples however, the behavior
was to default to DynamicPseudoType.

This change modifies the behavior so that dynamic types will be
speculatively converted to the returned unified type when a single type
can be found for a tuple or object.