function/stdlib: Don't panic on unknown tuple in formatlist
We had an exception to our short-circuit rule for tuple values so that we
would still get an opportunity to length-check them, but if they _do_ pass
the length check then we still can't iterate them.

Now we'll check the lengths of all of the arguments where the length is
known, even if an argument's value is unknown. This allows us to produce
an error message sooner for unknown tuples of incorrect length.