cty: Allow capsule types to overload some basic operations
So far this only includes the basic functionality that we expect all types
to have: GoString representations and equality tests. Right now it's not
a goal for capsule types to be able to support other operations or to
emulate built-in cty types because those other operations are
type-specific anyway and so we expect calling applications to already be
discriminating based on type when calling those and can thus call into
their own implementations directly if appropriate.

On the other hand, making these basic operations overridable means that
code that works generically with values of any type doesn't need to
re-implement all of the special behaviors for dealing with unknown values,
nulls, etc in order to properly emulate these generic operations for
application-specific types.