cty/function: Support for marks
Since most functions are not marks-aware, our default behavior is to
automatically unmark values coming in and then propagate the input marks
into the output, in similar vein to the built-in operations in cty.

However, functions may indicate that they are able to handle marks
internally, in which case they'll receive the marked values as-is and
are responsible for processing them correctly.