cty: Don't allow marked values in sets
Sets are tricky with marked values because marks don't contribute to the
value's definition of equality but yet if an unmarked and a marked value
that are equal were both written to a set then that would create a tricky
situation without a clear solution.

For now, marked values in sets are just blocked altogether. In a later
commit we'll improve this such that constructing a set with a marked value
applies the markings to the containing set instead.