cty: A set value inherits all the marks of its elements
Since a set can only contain each unique value once, but that value might
be added with multiple different marks, rather than trying to track the
marks for each distinct value separately we instead just propagate all of
the marks to the containing set.

This loses information, but creating a set _usually_ loses information,
and this compromise still ensures that any access to the values inside
must consider the marks of the set.