json, msgpack, gob: reject marked values
Marks can't survive serialization, so the caller must unmark the values
first.

In practice this would've been a panic anyway when trying to use the
integration methods to access the internals of the values, but this
up-front check allows it to be an error rather than a panic (consistent
with how we usually behave in these functions) and allows us to return
a PathError indicating where in the structure the mark is.