cty/function: Unmark values before function result type checking
Although this can potentially cause marks to be discarded in some
edge-cases, failing to do this forces all functions that try to read
values during their type checking to deal with the possibility of marked
values as well as unknown values, which is an unacceptable burden.

Instead, applications using marks must consider which functions they make
use of to ensure that the functions make sense within the use of marks.
For example, an application using marks to track that certain values come
from a sensitive source may wish to take care around both the JSONDecode
and Regex stdlib functions, because if the first argument to each were
marked as sensitive then the result type would disclose information about
that string without carrying the sensitive mark along with it.

This is consistent with the design goal that applications that are not
using marks should not have to worry about them, and that the burden of
dealing with marks falls on the applications electing to use them.