function/stdlib: Upstream various functions from HashiCorp Terraform
HashiCorp implemented additional functions for Terraform that previously
lived in the Terraform codebase. Now that they want to offer a similar
complement of functions in Packer, they are contributing some of the
Terraform functions upstream, where their design was general
enough to be considered as part of the "standard library".