a null dynamic value is a known value
When checking for nested dynamic types within complex types for equality
comparison, we need to also verify that those dynamic types are not
already known to be null.

Add the Value.HasDynamicValues method, to check for nested DynamicVal
values. HasDynamicValues indicates not only that this value is unknown,
but the type may change as well. This is used in the Value.Equals method
before checking the type for equality.